Completed production-ready styles on header
The header is responsive to mobile and larges-screen views. A darkmode/
lightmode button has been added with changeable state values but does
not change the theme of the webpage. A burger menu has been added with
changeable state values but does not open drawer. Context API was
implemented to manage UI state. When user Scrolls up, the header will
hide from view. When user scrolls down, the header will reappear.