Added basic structure to project/contact sections
The project section has border-radius on the images. The text field in
project section has padding for visual improvements. The Contact form is
created and utilizes components from MaterialUI for the text fields. The
button component was modified to change the button based on type "link"
and "submit" whivh affects Hero and Contact section